+++ app/src/main/res/layout/activity_sign_in.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:weightSum="2"
-    tools:context=".SignInActivity">
+    tools:context=".ui.user.SignInActivity">
 
     <ImageView
         android:layout_width="match_parent"
